+++ server/api/recipes/reviews/reviews.service.js
@@ -7,32 +7,57 @@ import {
   replaceReviews,
   getById as getByRecipeId,
 } from "../recipes.repository";
+import { InvalidAttribute, ResourceNotExists } from "../common.exception";
 
-let getReview = async (recipeId, data) => {
-  let user = await getByUserName(data.user);
-  data.userId = user._id;
-  data.recipeId = recipeId;
-  data.date = new Date();
-  delete data.user;
+const getReview = async (recipeId, data) => {
+  const user = await getByUserName(data.user);
+  if (user === null) {
+    throw new ResourceNotExists("User - " + data.user + " not exists");
+  } else {
+    data.userId = user._id;
+    data.recipeId = recipeId;
+    data.date = new Date();
+    delete data.user;
+  }
+};
+
+const validateReview = async (review) => {
+  if (!review.description) {
+    throw new InvalidAttribute("Review description required");
+  }
+  if (!review.rating) {
+    throw new InvalidAttribute("Review rating required");
+  }
+  if (!review.user) {
+    throw new InvalidAttribute("Review user required");
+  }
 };
 exports.create = async (recipeId, data) => {
+  await validateReview(data);
   await getReview(recipeId, data);
-  let reviewId = await create(data);
+  const reviewId = await create(data);
   await addReviewToRecipe(recipeId, reviewId);
 };
 
 exports.update = async (recipeId, reviewId, data) => {
+  await validateReview(data);
   await getReview(recipeId, data);
   await update(reviewId, data);
 };
 
 exports.deleteOne = async (recipeId, reviewId) => {
-  let recipe = await getByRecipeId(recipeId);
-  let reviews = recipe.reviews;
-  let reviewIndex = reviews.indexOf(reviewId);
-  if (reviewIndex !== -1) {
-    reviews.splice(reviewIndex, 1);
+  const recipe = await getByRecipeId(recipeId);
+  if (recipe === null) {
+    throw new ResourceNotExists("Recipe - " + recipeId + " not exists");
+  } else {
+    const reviews = recipe.reviews;
+    const reviewIndex = reviews.indexOf(reviewId);
+    if (reviewIndex !== -1) {
+      reviews.splice(reviewIndex, 1);
+    } else {
+      throw new ResourceNotExists("Review - " + reviewId + " not exists");
+    }
+    await replaceReviews(recipeId, reviews);
+    await deleteOne(reviewId);
   }
-  await replaceReviews(recipeId, reviews);
-  await deleteOne(reviewId);
 };
